Move application JS into head for turbolinks
Because of how turbolinks works,
having the application js in the body instead of the head is bad.
Not only was it in the wrong place, but the application js
looked to be called twice.

I commented out the jsapi and chartkick js because they aren't used.
If I'm missing something, that line may need to be reverted.
However, it did also fix a js type error :)